+++ python/md3qml/__init__.py
@@ -52,4 +52,4 @@ __all__ = [
     "version_string_c",
 ]
 
-__version__ = "1.1.2"
+__version__ = "1.1.3"